more cuda, not functional atm

+++ src/gpu-info.hh
@@ -9,5 +9,5 @@
 void print_device_props_short();
 void print_device_props_complete();
 int number_cuda_devices();
-int number_cuda_devices_minimum_mem_mb(int min_mem);
+int number_cuda_devices_minimum_mem_mb(size_t min_mem);
 int number_cuda_devices_minimum_cc(int min_cc);
